+++ test/column_metrics_test.dart
@@ -129,14 +129,14 @@ void main() {
       test('Empty', () {
         EasyTableModel model = EasyTableModel(columns: []);
         List<ColumnMetrics> list =
-            ColumnMetrics.resizable(model: model, dividerThickness: 10);
+            ColumnMetrics.resizable(model: model, maxWidth: 500, dividerThickness: 10);
         expect(list.length, 0);
       });
       test('Single - Divider', () {
         EasyTableModel model = EasyTableModel(
             columns: [EasyTableColumn(pinStatus: PinStatus.left, width: 50)]);
         List<ColumnMetrics> list =
-            ColumnMetrics.resizable(model: model, dividerThickness: 10);
+            ColumnMetrics.resizable(model: model, maxWidth: 500, dividerThickness: 10);
         expect(list.length, 1);
         expect(list[0].offset, 0);
         expect(list[0].width, 50);
@@ -146,7 +146,7 @@ void main() {
         EasyTableModel model = EasyTableModel(
             columns: [EasyTableColumn(pinStatus: PinStatus.left, width: 50)]);
         List<ColumnMetrics> list =
-            ColumnMetrics.resizable(model: model, dividerThickness: 10);
+            ColumnMetrics.resizable(model: model, maxWidth: 500, dividerThickness: 10);
         expect(list.length, 1);
         expect(list[0].offset, 0);
         expect(list[0].width, 50);
@@ -158,7 +158,7 @@ void main() {
           EasyTableColumn(pinStatus: PinStatus.none, width: 100)
         ]);
         List<ColumnMetrics> list =
-            ColumnMetrics.resizable(model: model, dividerThickness: 10);
+            ColumnMetrics.resizable(model: model, maxWidth: 500, dividerThickness: 10);
         expect(list.length, 2);
         expect(list[0].offset, 0);
         expect(list[0].width, 50);
@@ -173,7 +173,7 @@ void main() {
           EasyTableColumn(pinStatus: PinStatus.none, width: 100)
         ]);
         List<ColumnMetrics> list =
-            ColumnMetrics.resizable(model: model, dividerThickness: 0);
+            ColumnMetrics.resizable(model: model, maxWidth: 500, dividerThickness: 0);
         expect(list.length, 2);
         expect(list[0].offset, 0);
         expect(list[0].width, 50);